requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.36
xrpl-py==2.4.0
apscheduler==3.10.4
plyer==2.1.0
requests==2.32.3
pydantic[email]==2.9.2
structlog==24.1.0